+++ shop-product/product-service/src/main/java/com/mission/shop/product/service/cart/CartService.java
@@ -23,7 +23,7 @@ public interface CartService {
      * @param goodsId
      * @param buyNum
      */
-    public Cart addCartGoods(Long userId,Long goodsId,short buyNum);
+    public Cart addCartGoods(Long userId,Long goodsId,int buyNum);
 
     /**
      * @param userId
